requirements.txt
python==3.9
pytorch==1.13
pytorch-cuda==11.7
pytorch-lightning==1.8
pandas==1.4